@@ -3,7 +3,7 @@
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve --open",
+    "start": "ng serve --port 4400",
     "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
@@ -11,50 +11,54 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~7.0.0",
-    "@angular/cdk": "^7.0.1",
-    "@angular/common": "~7.0.0",
-    "@angular/compiler": "~7.0.0",
-    "@angular/core": "~7.0.0",
-    "@angular/forms": "~7.0.0",
-    "@angular/http": "~7.0.0",
-    "@angular/material": "^7.0.1",
-    "@angular/platform-browser": "~7.0.0",
-    "@angular/platform-browser-dynamic": "~7.0.0",
-    "@angular/router": "~7.0.0",
+    "@angular/animations": "~7.0.3",
+    "@angular/cdk": "^7.0.3",
+    "@angular/common": "~7.0.3",
+    "@angular/compiler": "~7.0.3",
+    "@angular/core": "~7.0.3",
+    "@angular/fire": "^5.1.0",
+    "@angular/forms": "~7.0.3",
+    "@angular/http": "~7.0.3",
+    "@angular/material": "^7.0.3",
+    "@angular/platform-browser": "~7.0.3",
+    "@angular/platform-browser-dynamic": "~7.0.3",
+    "@angular/router": "~7.0.3",
+    "angularfire2": "^5.1.0",
     "core-js": "^2.5.4",
-    "cors": "^2.8.4",
+    "cors": "^2.8.5",
     "express": "^4.16.4",
+    "firebase": "^5.5.8",
     "hammerjs": "^2.0.8",
-    "mongoose": "^5.3.6",
+    "mongoose": "^5.3.11",
     "passport": "^0.4.0",
     "passport-http": "^0.3.0",
     "passport-local": "^1.0.0",
-    "rxjs": "~6.3.3",
+    "rxjs": "^6.3.3",
+    "rxjs-compat": "^6.3.3",
     "zone.js": "~0.8.26"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.10.0",
-    "@angular/cli": "~7.0.2",
-    "@angular/compiler-cli": "~7.0.0",
-    "@angular/language-service": "~7.0.0",
-    "@types/jasmine": "~2.8.8",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
+    "@angular/cli": "~7.0.5",
+    "@angular/compiler-cli": "~7.0.3",
+    "@angular/language-service": "~7.0.3",
+    "@types/jasmine": "~2.8.11",
+    "@types/jasminewd2": "~2.0.6",
+    "@types/node": "~10.12.5",
     "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
+    "jasmine-core": "~3.3.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.0.0",
+    "karma": "~3.1.1",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "~2.0.1",
     "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "node-sass": "^4.9.4",
+    "karma-jasmine-html-reporter": "^1.4.0",
+    "node-sass": "^4.10.0",
     "protractor": "~5.4.0",
     "sass-loader": "^7.1.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": "~3.1.1",
-    "webpack": "^4.23.1"
+    "typescript": "~3.1.6",
+    "webpack": "^4.25.1"
   }
 }
