@@ -1,26 +1,23 @@
+import mongoose from 'mongoose';
 import passport from 'passport';
+import LocalStrategy from 'passport-local';
 
-const JwtStrategy = require('passport-jwt').Strategy,
-    ExtractJwt = require('passport-jwt').ExtractJwt;
+// Import database settings
+import './database';
+import '../models/user.model';
 
-// load up the user model
-const User = require('../models/user');
-const config = require('../config/database'); // get db config file
+const users = mongoose.model('users');
 
-module.exports = function(passport) {
-  const opts = {};
-  opts.jwtFromRequest = ExtractJwt.fromAuthHeaderWithScheme("jwt");
-  opts.secretOrKey = config.secret;
-  passport.use(new JwtStrategy(opts, function(jwt_payload, done) {
-    User.findOne({id: jwt_payload.id}, function(err, user) {
-          if (err) {
-              return done(err, false);
-          }
-          if (user) {
-              done(null, user);
-          } else {
-              done(null, false);
-          }
-      });
-  }));
-};
+passport.use(new LocalStrategy({
+  usernameField: 'user[email]',
+  passwordField: 'user[password]',
+}, (email, password, done) => {
+  users.findOne({ email })
+    .then((user) => {
+      if (!user || !user.validatePassword(password)) {
+        return done(null, false, { errors: { 'Email or password': 'is invalid' } });
+      }
+
+      return done(null, user);
+    }).catch(done);
+}));
