@@ -3,13 +3,21 @@ import mongoose from 'mongoose';
 const databasePort = '27017';
 const databaseName = 'fill-my-car';
 
+// Configure mongoose's promise to global promise
+mongoose.promise = global.Promise;
+mongoose.set('debug', true);
+
 mongoose.connect(`mongodb://localhost:${databasePort}/${databaseName}`, {
   useNewUrlParser: true,
   useCreateIndex: true
 });
 
 const connection = mongoose.connection;
+const consoleDivider = '\n\n******************************************\n';
+const connectionMessage = 'Mongo database is running';
 
 connection.once('open', () => {
-  console.log('MongoDB running!')
+  console.log(connectionMessage + consoleDivider)
 });
+
+module.exports = connection;
