@@ -16,23 +16,39 @@ import {
   MatSnackBarModule } from '@angular/material';
 import { AppComponent } from './app.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { VehicleService } from './vehicle.service';
+import { VehicleService } from './core/vehicle.service';
 import { HttpClientModule } from '@angular/common/http';
 import { HeaderComponent } from './components/header/header.component';
 import { FooterComponent } from './components/footer/footer.component';
-import { DashboardComponent } from './dashboard/dashboard.component';
-import { CreateComponent } from './create/create.component';
-import { EditComponent } from './edit/edit.component';
+import { DashboardPageComponent } from './pages/dashboard-page/dashboard-page.component';
+import { CreatePageComponent } from './pages/create-page/create-page.component';
+import { EditPageComponent } from './pages/edit-page/edit-page.component';
 import { RegisterComponent } from './components/register/register.component';
 import { LoginComponent } from './components/login/login.component';
 import { ContentComponent } from './components/content/content.component';
+import { UserPageComponent } from './pages/user-page/user-page.component';
+import { UserComponent } from './components/user/user.component';
+import { AngularFireModule } from 'angularfire2';
+import { AngularFirestoreModule } from 'angularfire2/firestore';
+import { AngularFireAuthModule } from 'angularfire2/auth';
+import { UserResolver } from '../app/components/user/user.resolver';
+import { AuthGuard } from './core/auth.guard';
+import { AuthService } from './core/auth.service';
+import { UserService } from './core/user.service';
+import { environment } from '../environments/environment';
+import { VehiclesPageComponent } from './pages/vehicles-page/vehicles-page.component';
+import { TripsPageComponent } from './pages/trips-page/trips-page.component';
+import { VehicleListComponent } from './components/vehicle-list/vehicle-list.component';
 
 const routes: Routes = [
-  { path: 'create', component: CreateComponent },
-  { path: 'edit/:id', component: EditComponent },
-  { path: 'dashboard', component: DashboardComponent },
-  { path: 'register', component: RegisterComponent },
-  { path: 'login', component: LoginComponent },
+  { path: 'create', component: CreatePageComponent },
+  { path: 'edit/:id', component: EditPageComponent },
+  { path: 'dashboard', component: DashboardPageComponent },
+  { path: 'register', component: RegisterComponent, canActivate: [AuthGuard] },
+  { path: 'login', component: LoginComponent, canActivate: [AuthGuard] },
+  { path: 'vehicles', component: VehiclesPageComponent, resolve: { data: UserResolver} },
+  { path: 'trips', component: TripsPageComponent, resolve: { data: UserResolver} },
+  { path: 'user', component: UserPageComponent, resolve: { data: UserResolver}},
   { path: '', redirectTo: 'dashboard', pathMatch: 'full' },
 ];
 
@@ -41,14 +57,23 @@ const routes: Routes = [
     AppComponent,
     HeaderComponent,
     FooterComponent,
-    DashboardComponent,
-    CreateComponent,
-    EditComponent,
+    DashboardPageComponent,
+    CreatePageComponent,
+    EditPageComponent,
     RegisterComponent,
     LoginComponent,
-    ContentComponent
+    TripsPageComponent,
+    UserComponent,
+    VehiclesPageComponent,
+    UserPageComponent,
+    ContentComponent,
+    TripsPageComponent,
+    VehicleListComponent
   ],
   imports: [
+    AngularFireModule.initializeApp(environment.firebase),
+    AngularFirestoreModule, // imports firebase/firestore, only needed for database features
+    AngularFireAuthModule, // imports firebase/auth, only needed for auth features
     BrowserModule,
     BrowserAnimationsModule,
     HttpClientModule,
@@ -66,7 +91,7 @@ const routes: Routes = [
     MatDividerModule,
     MatSnackBarModule
   ],
-  providers: [VehicleService],
+  providers: [VehicleService, AuthService, UserService, UserResolver, AuthGuard],
   bootstrap: [AppComponent]
 })
 
