@@ -1,15 +1,60 @@
 import { Component, OnInit } from '@angular/core';
+import { AuthService } from '../../core/auth.service';
+import { Router, Params } from '@angular/router';
+import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 
 @Component({
   selector: 'app-login',
   templateUrl: './login.component.html',
   styleUrls: ['./login.component.scss']
 })
-export class LoginComponent implements OnInit {
+export class LoginComponent {
 
-  constructor() { }
+  loginForm: FormGroup;
+  errorMessage: String = '';
 
-  ngOnInit() {
+  constructor(
+    public authService: AuthService,
+    private router: Router,
+    private fb: FormBuilder) {
+    this.createForm();
   }
 
+  createForm() {
+    this.loginForm = this.fb.group({
+      email: ['', Validators.required],
+      password: ['', Validators.required]
+    });
+  }
+
+  tryFacebookLogin() {
+    this.authService.doFacebookLogin()
+      .then(res => {
+        this.router.navigate(['/user']);
+      });
+  }
+
+  tryTwitterLogin() {
+    this.authService.doTwitterLogin()
+      .then(res => {
+        this.router.navigate(['/user']);
+      });
+  }
+
+  tryGoogleLogin() {
+    this.authService.doGoogleLogin()
+      .then(res => {
+        this.router.navigate(['/user']);
+      });
+  }
+
+  tryLogin(value) {
+    this.authService.doLogin(value)
+      .then(res => {
+        this.router.navigate(['/dashboard']);
+      }, err => {
+        console.log(err);
+        this.errorMessage = err.message;
+      });
+  }
 }
