@@ -1,3 +1,11 @@
 export const environment = {
-  production: true
+  production: true,
+  firebase: {
+    apiKey: 'AIzaSyA0D0sZDa4IGfdWBc_7TpQ84Vm3UCB3yvc',
+    authDomain: 'fill-my-car.firebaseapp.com',
+    databaseURL: 'https://fill-my-car.firebaseio.com',
+    projectId: 'fill-my-car',
+    storageBucket: 'fill-my-car.appspot.com',
+    messagingSenderId: '787032903752'
+  }
 };
