@@ -3,9 +3,16 @@
 // The list of file replacements can be found in `angular.json`.
 
 export const environment = {
-  production: false
+  production: false,
+  firebase: {
+    apiKey: 'AIzaSyA0D0sZDa4IGfdWBc_7TpQ84Vm3UCB3yvc',
+    authDomain: 'fill-my-car.firebaseapp.com',
+    databaseURL: 'https://fill-my-car.firebaseio.com',
+    projectId: 'fill-my-car',
+    storageBucket: 'fill-my-car.appspot.com',
+    messagingSenderId: '787032903752'
+  }
 };
-
 /*
  * For easier debugging in development mode, you can import the following file
  * to ignore zone related error stack frames such as `zone.run`, `zoneDelegate.invokeTask`.
